feat: split satellite services
