+++ packages/admin/src/index.ts
@@ -1,7 +1,15 @@
 export * from './services/mission-control.services';
 export * from './services/orbiter.services';
 export * from './services/releases.services';
-export * from './services/satellite.services';
+export * from './services/satellite.assets.services';
+export * from './services/satellite.config.services';
+export * from './services/satellite.controllers.services';
+export * from './services/satellite.docs.services';
+export * from './services/satellite.domains.services';
+export * from './services/satellite.memory.services';
+export * from './services/satellite.rules.services';
+export * from './services/satellite.upgrade.services';
+export * from './services/satellite.version.services';
 export type * from './types/actor.types';
 export type * from './types/build.types';
 export type * from './types/controllers.types';
